works so far... last modifief is a bit shitty on android

+++ core/src/main/java/son/sync/Receiver.java
@@ -21,20 +21,22 @@ public class Receiver {
                 System.out.println("File packet received");
                 var filePacket = (FilePacket)packet;
                 // var file = new File(syncFolder.folder, filePacket.getFilePath());
-                var syncFile = syncFolder.createSyncFile(filePacket.getFilePath());
+                var syncFile = syncFolder.getSyncFile(filePacket.getFilePath());
 
                 // createFolders(syncFolder.folder, filePacket.getFilePath());
                 if(filePacket.getSize() == 0) {
-                    syncFile.delete();
-                    System.out.println("File deleted");
+                    if(syncFile != null) syncFile.delete();
+                    System.out.println("File delete packet");
                 }
                 else {
+                    if(syncFile == null) 
+                        syncFile = syncFolder.createSyncFile(filePacket.getFilePath());
                     // endpoint.send(new BasePacket(PacketType.SEND_FILE));
                     System.out.println("Receiving file");
                     endpoint.receiveSyncFile(syncFile, filePacket.getSize());
                     System.out.println("File received");
                     // file.setLastModified(filePacket.getLastModified());
-                    syncFile.setLastModified(filePacket.getLastModified());
+                    // syncFile.setLastModified(filePacket.getLastModified());
                 }
                 endpoint.send(new BasePacket(PacketType.READY));
             }
